backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
httpx
pydantic
ollama>=0.4.0
humandelta>=0.1.0